ml/requirements.txt
xgboost==2.1.3
scikit-learn==1.6.1
pandas==2.2.3
numpy==1.26.4
optuna==4.2.0
shap==0.46.0
supabase==2.13.0
python-dotenv==1.1.0
fastapi==0.115.0
uvicorn==0.34.0
httpx==0.28.1
joblib==1.4.2